# openapi.yaml
openapi: 3.0.3
info:
  title: CEC HTTP Bridge API
  description: |
    REST API for controlling HDMI-CEC devices (Rust implementation). Power,
    volume, source/HDMI switching, remote keys, raw CEC commands, MQTT
    settings, a dev console surface, and real-time events (SSE + WebSocket).
    Intended for home automation, media centers, and scripting.

    ## Envelope
    Every JSON response uses the envelope
    `{"status":"success"|"error","message":"...","data":...}`.
    `message` and `data` are omitted when empty/absent. Long-running
    operations return `success` with `data.accepted=true` (HTTP 202).

    ## Authentication
    Auth is **optional** and active only when the server was started with a
    token (`-token <secret>` / `auth_token` in `config.json`). When active,
    every `/api` route **except `GET /api/health`** (and `GET /metrics`)
    requires the token via `Authorization: Bearer <token>`,
    `X-Auth-Token: <token>`, `?key=<token>`, or the `capi_token` cookie set by
    the UI login page. Mutating cookie-authenticated requests additionally
    require same-origin. When no token is configured, no auth is required.
  version: 2.0.0
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080
    description: Local (default)
  - url: http://{host}:8080
    description: Network
    variables:
      host:
        default: localhost
security:
  - BearerAuth: []
...
paths:
...
  /api/power/on:
    post:
      tags: [Power]
      summary: Power on TV
      description: Power on the TV (logical address 0).
      operationId: powerOn
      responses:
        '200':
          description: Power on command sent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/power/on/{address}:
    post:
      tags: [Power]
      summary: Power on device
      description: Power on a specific device by logical address.
      operationId: powerOnAddress
      parameters:
        - $ref: '#/components/parameters/LogicalAddress'
      responses:
        '200':
          description: Power on command sent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/power/off:
    post:
      tags: [Power]
      summary: Power off TV
      description: Put the TV (logical address 0) in standby.
      operationId: powerOff
      responses:
        '200':
          description: Standby command sent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/power/off/{address}:
    post:
      tags: [Power]
      summary: Power off device
      description: Put a specific device in standby.
      operationId: powerOffAddress
      parameters:
        - $ref: '#/components/parameters/LogicalAddress'
      responses:
        '200':
          description: Standby command sent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /api/volume/up:
    post:
      tags: [Volume]
      summary: Volume up
      description: >
        Volume up via libcec (targets the TV's system audio path). Use the
        addressed variant to target another device (e.g. 5 = audio system).
      operationId: volumeUp
      responses:
        '200':
          description: Volume up sent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/volume/up/{address}:
    post:
      tags: [Volume]
      summary: Volume up to specific device
      operationId: volumeUpAddress
      parameters:
        - $ref: '#/components/parameters/LogicalAddress'
      responses:
        '200':
          description: Volume up sent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/volume/down:
    post:
      tags: [Volume]
      summary: Volume down
      operationId: volumeDown
      responses:
        '200':
          description: Volume down sent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/volume/down/{address}:
    post:
      tags: [Volume]
      summary: Volume down to specific device
      operationId: volumeDownAddress
      parameters:
        - $ref: '#/components/parameters/LogicalAddress'
      responses:
        '200':
          description: Volume down sent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/volume/mute:
    post:
      tags: [Volume]
      summary: Toggle mute
      operationId: volumeMute
      responses:
        '200':
          description: Mute toggle sent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/volume/mute/{address}:
    post:
      tags: [Volume]
      summary: Toggle mute on specific device
      operationId: volumeMuteAddress
      parameters:
        - $ref: '#/components/parameters/LogicalAddress'
      responses:
        '200':
          description: Mute toggle sent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'

  # ────────────────────────────── Source / HDMI ───────────────────────────
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: >
        Optional. Active only when the server runs with a token
        (`-token <secret>`). `X-Auth-Token`, `?key=`, and the `capi_token`
        cookie are accepted equivalently. When no token is configured this
        security requirement is a no-op.
  parameters:
    LogicalAddress:
      name: address
      in: path
      required: true
      schema:
        type: integer
        minimum: 0
        maximum: 15
      description: CEC logical address (0-15; 15 is the broadcast address)
  schemas:
    ApiResponse:
      type: object
      description: >
        Standard envelope. `message` omitted when empty; `data` omitted when
        absent. Async acceptance returns success with data.accepted=true.
      required: [status]
      properties:
        status:
          type: string
          enum: [success, error]
        message:
          type: string
          description: Human-readable message (omitted when empty)
        data:
          description: Response payload (varies by endpoint; omitted when absent)
          nullable: true
...
  responses:
    BadRequest:
      description: Invalid parameters or request body
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponse'
          example:
            status: error
            message: Invalid logical address
    InternalError:
      description: CEC operation failed or server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponse'
          example:
            status: error
            message: failed to open adapter